Update fh.js
New method additions; improvements

+++ fh.js
@@ -6,23 +6,29 @@
 */
 
 var fh = {
-	api_url : 'http://10.66.228.153/tars/api/web/v1/',
-	version :  1.1,
+	api_url : 'http://10.66.228.153/project/api/',
+	//api_url : 'http://ncloudhost.com/api/',
+	version :  1.3,
+	app_name : 'App Name',
 	debug : true,
 	ERR : {
 		NET : {
 			GENERIC : 'Please check your internet connection.',
-			UNEXPECTED: 'An unexpected error occured.',
+			UNEXPECTED : 'An unexpected error occurred.',
 		},
 		DATA : {
 			LOADING: 'We\'re loading up for you.',
 		},
+		SESSION : {
+			INVALID : 'Invalid session. Please try again.',
+		}
+
 	},
 
 	
 	alert : function(m){
 
-		navigator.notification.alert(m, null, 'Message');
+		navigator.notification.alert(m, null, this.app_name + ' - Message');
 	},
 	
 
@@ -33,6 +39,12 @@ var fh = {
 		setTimeout(function(){ $.mobile.loading('hide');
 			if(callback && typeof callback == "function")  callback();
 		}, delay);
+	},
+
+	confirm : function(message)
+	{
+		cnf =  window.confirm(message);
+		return cnf;
 	},
 
 	
@@ -58,7 +70,9 @@ var fh = {
 					break;
 
 				case '@string':
-					rs = (isNaN(rs)) ? rs : false;
+					rs = rs; // not necessary; just to understand
+					// FIX HERE @todo
+					//(isNaN(rs)) ? rs : false;
 					break;
 
 				case '@date':
@@ -74,7 +88,8 @@ var fh = {
 					break;
 
 				case '@mobile':
-					//// Indian standard mobile pattern 09987724488 +919987724488 0919987724488 are valid numbers
+					//// Indian standard mobile pattern
+					///  9987724488 ^ 09987724488 ^ +919987724488 ^ 0919987724488 are valid numbers
 					var re = /^([0|\+[0-9]{1,3})?([7-9][0-9]{9})$/;
 					rs = (re.test(rs)) ? rs : false;
 					break;
@@ -119,12 +134,16 @@ var fh = {
 		if(fetch) return fetch;
 		else return false;
 	},
-	
 
-	confirm : function(message)
+	rmLocal : function (k)
 	{
-		cnf =  window.confirm(message);
-		return cnf;
+		if(typeof k === 'object')
+		{
+			Object.keys(k).forEach(function (key) {
+				window.localStorage.removeItem(k[key]);
+			});
+		}
+		else window.localStorage.removeItem(k);
 	},
 
 
@@ -152,9 +171,23 @@ var fh = {
 
 
 	logout : function (holdRedirect) {
-		localStorage.clear();
+
 		if(!holdRedirect)
-			this.changePage('login.html');
+		{
+			navigator.notification.confirm('Are you sure to logout ?', (function(idx){
+				if(idx === 1)
+				{
+					localStorage.clear();
+					fh.changePage('login.html');
+				}
+			}), this.app_name + ' - Confirm', ['Logout', 'Cancel']);
+		}
+		else
+		{
+			localStorage.clear();
+			fh.changePage('login.html');
+		}
+
 	},
 	
 
@@ -211,12 +244,48 @@ var fh = {
 		var logged = false;
 		if(!this.getLocal('uid') || !this.getLocal('email'))
 		{
+			log('Auth forcing logout from - ' + fh.curPageName());
 			if(holdRedirect) this.logout(true);
 			else this.logout();
 		}
 		else var logged = true;
 		////
 		return logged;
+	},
+
+	setSessionAuth : function(_p, token)
+	{
+		if(_p && token)
+		{
+			var token_param = _p + '_token';
+			var time_param = _p  + '_token_time';
+			_o = {};
+			_o[token_param] = token;
+			_o[time_param] = new Date();
+			fh.setLocal(_o);
+		}
+		else return false;
+
+	},
+
+	getSessionAuth : function(_p, token)
+	{
+		if(_p && token)
+		{
+			var token_param = _p + '_token';
+			var time_param = _p  + '_token_time';
+			var req_token = fh.getLocal(token_param);
+			var req_time = (fh.getLocal(time_param)) ? new Date(fh.getLocal(time_param)) : false;
+			var cur_time = new Date();
+			var diff = Math.abs(cur_time - req_time);
+			var diff_min = Math.floor((diff/1000)/60);
+			log('Diff min : ' + diff_min);
+			if(req_token && req_time  && req_token == token && diff_min <= 1)
+			return true;
+			else return false;
+		}
+		else return false;
+
 	},
 	
 
@@ -236,9 +305,10 @@ var fh = {
 	{
 		navigator.notification.confirm('Are sure to exit the application ?', (function(idx){
 			if(idx === 1) navigator.app.exitApp();
-		}), 'Confirm', ['Exit', 'Cancel']);
+		}), this.app_name + ' - Confirm', ['Exit', 'Cancel']);
 
 	},
+
 
 
 	randString : function(len)
@@ -273,6 +343,33 @@ var fh = {
 
 
 };
+
+
+Array.prototype.inArray = function(comparer) {
+	for(var i=0; i < this.length; i++) {
+		if(comparer(this[i])) return true;
+	}
+	return false;
+};
+
+Array.prototype.pushIfNotExist = function(element, comparer) {
+	if (!this.inArray(comparer)) {
+		this.push(element);
+	}
+};
+
+
+Array.prototype.cleanPush = function(elem, key) {
+	for(var i=0; i < this.length; i++) {
+		if(typeof this[i]['qid'] != 'undefined' && this[i]['qid'] == key)
+		{
+			this.splice(i, 1);
+		}
+	}
+	this.push(elem);  
+};
+
+
